tests/wpt/tests/css/css-sizing/responsive-iframe/parsing/frame-sizing-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic Size Test: `frame-sizing`</title>
<link rel="author" title="Koji Ishii" href="mailto:kojii@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#responsive-iframes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id=target></div>
<script>
test_computed_value("frame-sizing", "auto");

test_computed_value("frame-sizing", "content-width");
test_computed_value("frame-sizing", "content-height");
test_computed_value("frame-sizing", "content-inline-size");
test_computed_value("frame-sizing", "content-block-size");
</script>
</body>
</html>
